[shweta] added the float code

+++ GenricExamples/Program.cs
@@ -11,13 +11,13 @@ namespace GenricExamples
             
             Console.WriteLine("Hello World!");
             Console.WriteLine("Enter the First number:");
-            int f = Convert.ToInt32(Console.ReadLine());
+            float f = Convert.ToSingle(Console.ReadLine());
 
             Console.WriteLine("Enter the second number:");
-            int s = Convert.ToInt32(Console.ReadLine());
+            float s = Convert.ToSingle(Console.ReadLine());
 
             Console.WriteLine("Enter the third number:");
-            int t = Convert.ToInt32(Console.ReadLine());
+            float t = Convert.ToSingle(Console.ReadLine());
 
             SimpleClass simple = new SimpleClass(f,s,t);
             simple.findMax();
